+++ laktory/models/pipeline/pipelinechild.py
@@ -88,7 +88,7 @@ class PipelineChild(BaseModel):
     @property
     def parent_pipeline_node(self):
 
-        from laktory.models.pipelinenode import PipelineNode
+        from laktory.models.pipeline.pipelinenode import PipelineNode
 
         def _get_pl(o):
             parent = getattr(o, "_parent", None)