+++ tests/test_job.py
@@ -14,7 +14,7 @@ job = Job(
     tasks=[
         {
             "depends_ons": [{"task_key": "ingestion"}],
-            "pipeline_task": {"pipeline_id": "${resources.pl-spark-dlt.id}"},
+            "pipeline_task": {"pipeline_id": "${resources.dlt-pipeline-pl-spark-dlt.id}"},
             "task_key": "pipeline",
         },
         {
@@ -159,7 +159,7 @@ def test_job_model():
                 "notebook_task": None,
                 "notification_settings": None,
                 "pipeline_task": {
-                    "pipeline_id": "${resources.pl-spark-dlt.id}",
+                    "pipeline_id": "${resources.dlt-pipeline-pl-spark-dlt.id}",
                     "full_refresh": None,
                 },
                 "retry_on_timeout": None,
@@ -318,7 +318,7 @@ def test_job_pulumi():
             },
             {
                 "depends_ons": [{"task_key": "ingestion"}],
-                "pipeline_task": {"pipeline_id": "${resources.pl-spark-dlt.id}"},
+                "pipeline_task": {"pipeline_id": "${resources.dlt-pipeline-pl-spark-dlt.id}"},
                 "task_key": "pipeline",
             },
             {